core: Move errors to variables
Some of the error strings are duplicated and have slightly different
wording or casing.

This moves the majority of the errors to variables so they can be
reused.

The errVaultNotCreated error also clarifies that the vault could be
missing; if an existing Red October instance is started but the DB file
is missing (such as on stateless systems) the existing error message can
be confusing.